Fix update npm packages with critical vulnerabilities
Githib security updates (and npm audit) flagged two packages as
having critical vulnerabilities. Whilst these are not used in
a user-facing context, they are used as part of the build process
and execute code on developer and CI environments. Therefore they
should be patched.

`minimist` is used as a grandchild transitive dependency of
`gulp-sass-lint`. Unfortunately the intermediary depencency has not
updated from the vulnerable version, but we can override this. As
this dependency is only used as part of build/dev tooling, it's
unlikely to cause an issue. This has been pinned to a
non-vulnerable  version.

`lodash.template` is a transitive dependency of `gulp-css-url-adjuster`.
This has been overridden to use a non-vulnerable version.