Set `SameSite = Lax` on our 4 cookies
Read more about the `SameSite` value at
https://web.dev/samesite-cookies-explained/

Most browsers already default to `SameSite = 'Lax'` so
the behaviour added in this commit is already happening for most
of our users. However a small number will be on older browsers
that still have a default of `SameSite = 'None'`.

We think it is better to be explicit rather than implicit on the
value of `SameSite` rather than using browser defaults.

We think it is better to be restrictive and not allow `SameSite = None`
as we can't think of a valid use case for cross site requests to send
our cookies with their requests. This is particularly important for the
`notify_admin_session` cookie as it is particularly sensitive. Note, we
already block people embedding Notify in iframes, but there could
be people doing things like embedding our images on their websites but
these have no need to send our cookies with those requests.

We think that we don't want to set `SameSite = Strict` as this would
stop people opening links to Notify (say from their email client) that
would be logged in or remember their cookie preferences. It would
be a bad user experience.