Add new pattern for presentation code
This demos the new pattern for the block of related code that we
need to render the page to copy a template - formerly this was the
(quite confusing) UserTemplateLists class.

Since this is a new pattern, we need to decide how to name the
classes. We could call them all "XYZPresenter", but that would be
quite verbose. I've decided to use the package name to make this
clear instead i.e. presenters.XYZ. This isn't something we do very
much elsewhere so could be confusing, especially if a future dev
decided to import directly from a module.