Change how we show empty folders to users
If a user is filtering by a specific template type, then it always
makes sense to say a folder doesn't have any templates of that type
in it. Otherwise, we can just say the folder is empty.

I don't think there's any value in saying the folder is empty in the
edge case where a user is filtering by a specific template type and
the folder has no templates at all in it.