Fix back link on final page to copy a template
This was broken in two ways:

- It would take the user to a 404 page if the template they copy is
from a different service. To reproduce:

- In all other cases it would exit the copy workflow and go to the
preview page for the original template.

Steps to reproduce:

- Make two services: A and B. Make a template in service A.

- Switch to service B and copy the template from service A but stop
at the last page, where you see the draft template.

- Click back and see how the URL is /services/B/templates/<id> and
not /services/A/templates/<id>.

Even if the URL didn't 404, we would still be forcing the user to
start the copy workflow all over again.

This fixes that and adds a couple more tests for the specific back
links in the copy workflow.