Add a label to the button for errors if present
The existing implementation shows errors if they
return from the file selection process. This
version hides the label that contains them and so
hides the errors. This is unintentional.

Buttons don't need labels by default so this adds
one, to contain the errors, only if errors exist
and shows it to match the existing visual design.