Replace table on user-profile page with summary-list component
The '/user-profile' page was using a table to display the user data, and
this commit changes it to use the GOV.UK Frontend summary-list component
instead. The component takes rows as an input and it's not possible to
use conditional logic inside the component. This means that we need to
build all the data for the rows in variables first, then pass it to the
component.

It's not possible to pass ids to the component, but the existing ids on
the page were only being used in tests, so I've changed how we find rows
in the tests.