set email auth permission for new services created by email auth users
if someone signs in with email auth, then we know that they already
belong to a service where this setting was enabled. That means someone
already confirmed they sign in with a smart card, etc. It's reasonable
to assume that a new team that that user is creating will also have the
same restrictions in place, so lets let them set auth type for any users
they invite right from the off