Show the branding in the email branding pool on org settings page
This adds a new row the settings table on the organisation page called
'Email branding options'. The row displays the names of all email
brandings in the branding pool for that organisation, or 'None' if the
branding pool is empty. There's no 'change' link since we haven't
implemented anything which lets you change the branding pool yet.