Fix case where NHS branding can appear twice on ChooseEmailBrandingForm
NHS branding gets added to the `ChooseEmailBrandingForm` as a special
case (if a service's organisation is an NHS org and the service doesn't
already have NHS branding). However, if this condition applies and NHS
branding is also in the branding pool it will appear twice, with both
options behaving differently. The first option will take you to the
`.email_branding_nhs` route, while the second will take you to the
`.email_branding_pool_option` route.

This change deduplicates the options on the forms to only show the
"special NHS option" (the one that redirects to `.email_branding_nhs`.