Add new page to change organisation email branding
This adds a new platform admin page,
`.organisation_email_branding_options`, which is linked to from the
organisation settings page. The page shows a list of checkboxes for all
the available email branding - branding that is in the branding pool for
the organisation is checked.

The functionality to submit the form will be added later.