Remove styles for nested bullet points
We no longer have any nested bullet points in the code (they used to
exist on the pricing page) so this removes their style. If we do want
them in future we would need to rewrite these styles anyway since we're
removing GOV.UK Elements, which contains the `.list-bullet` styles.