Move GOVUK Elements forms styles into app
We still use these styles in several places and
replacing them will involve moving that code to
GOVUK Frontend. That work is already captured in
stories so this moves the styles into our app for
now, so as not to block the rest of the work
removing the GOVUK Elements library.