Copy GOVUK Elements typography styles to this repo
This copies the typography styles from GOVUK Elements to this repo so
that we can remove some more of the dependency on GOVUK Elements. It
deletes styles in the file that we weren't using and keeps those that we
still are - we will need to create separate stories to remove the styles
that are still in use.