Make form to set email branding pool for an org work
We had a form to let you pick email branding to add to the branding
pool for an org, but the functionality had not been implemented. This
adds the button and the POST endpoint for the form. At the moment, we
redirect to the organisation settings page when the form is successfully
submitted. Once the page to preview all email brandings in the pool for
an org exists, we can update the code to redirect there instead.