Support API errors for `email_from`
We are updating the notifications-api to reject `email_from` values
which contain 'uncleanable' characters (basically anything that can't be
reduced to alphanumeric characters or full-stops).

This is because AWS SES doesn't allow sending rich unicode characters in
the local portion of an email address. We could punycode the values, but
this looks very odd in emails and so for now we're leaving this alone -
to be addressed potentially in the future. We're also leaving it as a
future endeavour to allow separating service name and email_from, due to
the need for research+design.