Make new profile page align with grid
Our grid on these page is 4 columns, so 25% of the page per column.

This matches the width of the service navigation, on pages which have
it.

I’ve added some extra CSS so that when we apply this pattern to stuff
like the service settings page it should also maintain the grid.