display default branding and all the available email branding pool options and their previews
wrote organisation_email_branding_options view function to display email branding pool options and their previews
wrote a few unit testsorganisation_email_branding_options

linked add other options button to add email branding options page
In order to link the Add other options button to the add email branding options page, a few changes had to be
made to the page_footer macro.

There is now the option of a secondary button and if the secondary button is just to be a button linking to another page
as is the case in this instance.

The add_organisation_email_branding_options function was updated to point to the email-branding-options page where all
the previews of the options are displayed.