Add date for remaining issues
This is based on all remaining work being tackled
sequentially and implemented by one person, due to
these issues being so closely related to the
experience of parts of our user interface when
used with assistive tech' (which only I have the
ability to test and work with on the team).

We expect me to become available by the start of
December. The following story is majority content
design so we plan to play earlier:

'The way we label the different regions of our
pages is confusing for screen reader users'

I've timed and arranged the other stories to be
played in this order:

1. 'The JAWS screen reader announces template
folder page headings as multiple sentences'
2. 'The JAWS and NVDA screen readers do not
announce the ‘copy to clipboard’ button.'
3. 'Some icons do not appear when using high
contrast mode on Windows.'
4. 'Choosing the date and time you want to send a
message using a screen reader can be confusing.'
5. 'Expanding and collapsing groups of checkboxes
using a screen reader can be confusing'

Timings for each are based on how well we
understand them and the complexity of potential
solutions.