Add police force areas in England and Wales
In some areas its more intuitive to select the area covered by the
police force, rather than a local authority boundary.

This is often the case with cities, for example:
- Greater Manchester can’t be directly selected from our library,
  instead you need to add City of Manchester, Salford, etc.
- London can’t be directly selected from our library, instead you need
  to add Hackney, Tower Hamlets, etc.

The ONS publishes the boundaries for police force areas. They don’t have
multiple levels or hierarchies, so we can import them in much the same
way as we do for countries.

This also has the benefit of having a library of areas which aligns with
the areas covered by each Local Resilience Forum – they are the same.