Cache populations for police force areas
There isn’t directly-published population data for police force areas,
as far as I can tell.

We use population data to work out roughly how many phones are in a
given area.

We could modify the app to calculate the estimated population of an area
with missing data on the fly, like we do for custom areas[1]. However this
is a bit slower, so to keep the app feeling snappy this commit adds some
hard coded population numbers for each of the police force areas.

1. In other words by summing the populations of all the electoral wards
   which overlap the areas