Merge Metropolitan and City of London Police areas
The one exception to the Local Resilience Forum and police force
boundaries being the same is in London, where the London Resilience
Forum covers both the Metropolitan Police area and the City of London
Police area.

As requested, this commit merges these two areas into one. This should
stop someone picking ‘Metropolitan’ and accidentally leaving a hole in
the middle for the City of London (it’s easy to miss this on the map).

If someone does specifically want to target the City of London they can
do so by picking it from the local authorities library.