Update local authority geodata to 2021
The data we are using for the electoral wards is from 2020. The data we
are using for local authorities is from 2019.

This means that local authorities created in 2020 would be missing from
our dataset (or we would would have their predecessors) so it was not
possible to match their electoral wards to the the correct parent. This
meant that the local authority appeared not to have any electoral wards.

The fix for this is to update our datasets the 2021 releases, which are
the latest. Importantly this means that all the local authority datasets
are now from the same year, thereby fixing the mismatch.

Using outdated datasets meant that local authorities changed in 2020 and
2021 didn’t exist at all in our database. Now they do, including
examples like:
- Buckinghamshire (a unitary authority created in April 2020 from the
  merger of South Bucks, Chiltern, Wycombe and Aylesbury Vale in
  accordance with the The Buckinghamshire (Structural Changes) Order
  2019[1])
- Hartlepool (whose ward boundaries were redrawn in May 2021 in
  accordance with The Hartlepool (Electoral Changes) Order 2019[2])

Old alerts will still work even though they have been removed from the
database. They will use the polygons stored in the API instead, and be
treated as custom areas.

1. https://www.legislation.gov.uk/ukdsi/2019/9780111183717/contents
2. https://www.legislation.gov.uk/uksi/2019/1089/contents/made