Move GOVUK Elements reset into the app CSS
The GOVUK Elements reset sits before everything
else in our cascade and targets elements directly
so is too dangerous to remove. This instead moves
it into our stylesheet as a reset for the app.

Note: the position of the new `reset.scss` file is
higher in the cascade than previously, importantly
above `globals.scss`.

We analysed the contents of both files and the
only styles that could overlap are those for
`strong` and `b` tags. All uses of these tags in
our templates include the
`govuk-!-font-weight-bold` class which overrides
all other styles so we believe this overlap is
acceptable.

Note: all GOVUK Elements code we've added to our
CSS so far has gone in the govuk_elements folder.
The intention of this was to make it clear the
origin of this code and that it would be removed
in time. This reset CSS should stay in the app so
is not included in that folder.