Fix govuk_elements folder and its import path
Makes the govuk_elements folder like the
govuk-frontend folder (using a hyphen instead of
an underscore).

Also rewrites their import paths in main.scss to
make it more obvious that the path refers to a
local folder. Again, this is also to match the
existing path for the govuk-frontend folder.