Clear organisation email branding pool
When we set a new default email branding for an organisation, it gets
added to their email branding pool, so we need to clear that cache
entry.

This is a slightly weird way of doing - we have to 'know' that the API
call does this extra thing in this situation and manage the cache from
the frontend. But that's how it works at the moment, so this is at least
consistent...