Rename tables.scss file that was copied from GOV.UK Elements
We copied the file `tables.sccs` from GOV.UK Elements into our code as a
temporary fix until we no longer need it. This changes its name to
`_tables.scss` to match the naming of the other files we've copied
across from Elements and to ensure that it doesn't get generated into a
CSS file.