Bump utils from 56.0.2 to 56.0.3
This is a patch version bump to bring in an updated CSS class name. We
no longer have the styles for `.visually-hidden` that the letter preview
template uses - we removed them in https://github.com/alphagov/notifications-admin/pull/4326,
so this updates utils to a version with the updated class name.