Refactor window.location mock
Now uses URL() under the hood.

The existing method used regexps that were
incorrect. This simplifies how it works to just
proxy all calls to parts of the url to an instance
of URL, to make it do all the hard work.

Also removes url argument.

The resulting mock should start as a wrapper
around the current URL with modifications through
props, to match how window.location is used.