Update tests that use timers
Jest 26 changed how it managed faking time by
re-implementing its system with a wrapper wround
sinonjs/fake-timers, which was set as the default
method in Jest 27:

https://jestjs.io/blog/2020/05/05/jest-26#new-fake-timers

It doesn't change any parts of the API we already
use but now mocks Date so we can't use
jest-date-mock any more. That's ok though, the new
implementation lets you do the same through the
Jest API:

https://jestjs.io/docs/jest-object#jestsetsystemtimenow-number--date

This also removes the jest-date-mock package as it
is no longer needed. It also removes an import for
it from updateStatus.test.js which is redundant.