Add IDs to headings on the performance page
I often try to link to these in support tickets but am not able to
link directly to the section I'd like to.

I think these would be useful.

The ids have not been written or reviewed by a content designer. I
am not sure whether they need to or not.