Fix tracking of form submissions
JSDOM doesn't implement form submissions but the
tests for preventDuplicateFormSubmissions need to
track them.

Originally, we were able to do this by mocking the
submit method but JSDOM's form internals changed
when this pull request was merged which stopped
this working:

https://github.com/jsdom/jsdom/pull/3249

Thinking about it, I decided because JSDOM outputs
a very specific error for form submissions,
tracking that would be simpler than trying to mock
their internal code. This changes the test to do
that instead, removing the need for the
formSubmitSpy helper.