Move toolkit JS into this codebase
We intend to keep using the module pattern in
modules.js so this adds it to our JS folder.

We intend to remove show-hide-content.js but still
have radios using the GOVUK Elements HTML & CSS
that rely on it. This follows the approach we took
with GOVUK Elements code of moving it into the
codebase for now and deleting it when the GOVUK
Elements code that uses it is removed.

Note: this also changes the
app/assets/javascripts/modules folder to
app/assets/javascripts/esm to make it clear that
its contents (ECMAScript modules) are different to
the code in modules.js.