Refactor toolkit JS to pass local linting
The toolkit JS was refactored to pass StandardJS
linting rules:

https://github.com/alphagov/govuk_frontend_toolkit/pull/334

We haven't moved to them yet and still lint
according to JSHint + our own config (see
.jshintrc). This refactors the toolkit JS to pass
our rules.

Note: we will move to StandardJS rules in the
future but doing so requires changing all our JS
and tests so cannot be included in this story.