Fix problematic data
There are 3 polygons which I’ve had to manually edit in the source data
to avoid invalid topologies in our library.

Bathavon South
---

This polygon has a hole in it which needed opening out to preserve a
similar area between the simplified and source polygons, as we had to
with its previous version in 779ac74:

> This is the only way I can think to stop this shape self-intersecting
> without drastically changing its area (i.e. filling the hole in the
> donut).
>
> This is the only area in our library which is a genuine donut and
> presents this problem

Tolleshunt D’Arcy
---

This polygon has the same problem as Bathavon South, where it contains
a hole which is not defined as its own shape but by the meeting of two
edges at zero distance.

Oakham South
---

This one had two points from separate polygons very close together which
caused an artefact when merging into a single polygon during our
simplification process.