Move GOV.UK Frontend above our copy of GOV.UK Elements in SCSS cascade
We're going to change some of the Elements code we copied into this app
to use GOV.UK Frontend typography instead of the GOV.UK Frontend Toolkit
typography. This means that we need to import GOV.UK Frontend before the
GOV.UK Elements files, so this change the order that we import things in
main.scss.