Add error handling for email branding name
We had some 500s from the API this morning when a platform admin was
trying to add an email branding with a name that was already in use.
This raised a 500 on notifications-admin, and the platform admin didn't
know what was going wrong.

Let's try to handle this a bit more gracefully.