Don’t allow @ symbol in organisation domains
The domain will only work if it’s just the domain. If it’s an email
address or part thereof it won’t be accepted.

We have a few of these in our production database, let’s not have any
more.

We can’t put these errors directly on the fields, because they will be
replaced by the code in `list-entry.js`, which doesn’t account for
error messages when re-rendering the fields. Instead we show an error
banner at the top of the page, same as we do for the guest list page.