Update the 4 shades of grey
We have `$govuk-compatibility-govukfrontendtoolkit` set to `true`, which
means that we still have four shades of grey available to use. When we
turn legacy mode off, we will only have three shades of grey so we will
need to make further changes at that point. For now, everything looks
the same.