Replace toolkit tabular font variable
Replaces a Sass variable representing the list of
font families used for the tabular cut of the NTA
font (the one used before GOVUK Frontend v3).

This uses the `govuk-font` helper because the new
font (used in GOVUK Frontend after v3) contains
the tabular numbers cut instead of it being a
separate font family so we can't set it by
switching families.