Use quotes to disambiguate
Names of services might not always read neatly in a sentence. By adding
quotes around them we make it clear which words are part of the name of
the service or branding we are referring to.

I don’t think we need to do this for organisations; their names tend to
be more noun-based so have less need for disambiguation.

Also removes one place where we are repeating the name of the branding
without needing to; referring to it with ‘this’ should be clear enough
and makes the page a bit cleaner