Stop importing typography from GOV.UK Frontend Toolkit
We were relying on the `device-pixel-ratio` mixin that the typography file
from Toolkit imported. Since we no longer have access to the mixin, we
could import the "_device-pixels.scss" file it's in instead but there's
a version in GOV.UK Frontend we can use which is identical so we now use
that.