Ensure that changing service branding to GOV.UK does not 404
When changing service branding to GOV.UK, we are actually changing
it to no branding, as GOV.UK is what we show as a default when
a service or an rganisation has no branding set.

When we introduced a page confirming if new service branding
should be added to the pool, we introduced a bug where the view
was trying to find a branding with the id of __NONE__.

This commit fixes that bug.