Re-add page to confirm NHS branding
In 92f76638c891487889b7debcdc8c7914da5a804a, when we let people choose
NHS branding for themselves, we added a bespoke confirmation page with
guidance about whether using the NHS branding was appropriate.

This was lost once we put NHS branding in the pool, and choosing it was
just picking another `id`, rather than `nhs` as a special case.

This commit restores the behaviour by adding an extra check against the
`id` of the NHS branding, and treating it the same as if they had
previously chosen `nhs` as a special case.