Remove ‘magic’ NHS branding choice
We had a special branding with the ID `nhs` which we could give users
as a choice even if it wasn’t in their pool.

Once we made the NHS branding into a real branding like any other, this
meant we had to de-duplicate the magic option and the real option.

This commit removes the magic option and uses the real option
everywhere, to make the code simpler.