Allow fingerprint or face recognition security key
Modern browsers accept two type of authenticator (security key)
attachment:
1. `cross-platform`: the authenticator (security key) can be removed
   (unplugged) from the platform (your laptop) and used on another
   platform (laptop). In other words a USB stick.
2. `platform`: the authenticator (security key) is built into the
   platform (your laptop)

An example of a platform authenticator is the touch ID sensor on newer
Macbooks. It stores a hash of a scan of a fingerprint, which is then
checked when a user wants to authenticate.

This commit sets the attachment to `None`, which means that both
platform and cross-platform authenticators can be added.

This means we can use touch ID to log into Notify, which is useful when
you leave your Yubikey at home on a mandatory day in the office.

See:
- W3C standard: https://www.w3.org/TR/webauthn-2/#dom-authenticatorselectioncriteria-authenticatorattachment
- Implementation in the library we are using: https://github.com/Yubico/python-fido2/blob/8c00d0494501028135fd13adbe8c56a8d8b7e437/fido2/server.py#L176-L177