Add email address to events when archiving a user
We now store the email address of the user being archived in the events
table as part of the `archive_user` event.

This will enable us, in a separate story, to delete the user's name and
email from their row in the user table when we archive them and not
have to introduce a new job to remove user data that is older than 1
year for archived users.