Replace use of tint with shade for area list label
In 4ed653c23763e673d889506d1a97aa7acd93c597 we replaced `mix` with
`govuk-tint`.

`mix`, when given a colour and black, makes the colour darker.

`tint` makes a colour lighter.

This commit replaces the use of `tint` with `shade`. `shade` darkens a
colour, so is more similar to mixing it with black.

This means the label text maintains good colour contrast with its
background.

Colour values:

| `mix(black, 66%)` | `tint(33%)` | `shade(66%)` | shade(60%) |
| ----------------- | ----------- | ------------ | ---------- |
| #072840           | #5493c3     | #002038      | #002642    |