Re-add image assets for examples of branding
These were removed in https://github.com/alphagov/notifications-admin/pull/3123/files

This commit adds some of them back, so we can use them in future email
branding journeys.

`single_identity.png` is new and comes from the prototype – it’s not
an option we offered before.

Feel free to merge this as-is or cherry-pick the commit onto a different
branch.