Make helper methods for pool excluding default
There’s a few places where we’re trying to get a list of all the
branding in the pool, _except_ for a special case liek the organisation
default, or the current services branding.

We also acheive this in different ways in different places, for example
using `filter` or using list comprehensions.

This commit makes a helper method on the branding model which unifies
this logic into one place, and means fewer gnarly one liners in the
view layer.

This required changing some test data to ensure that it is out-of-order
so that the sorting of the checkboxes is covered by the test.