Add images and coloured stripes to form
This commit adds a preview of the logo and the colour to each option in
the form, to make it easier to see what you are picking.

The implementation is a bit specific, but think that’s OK since we’re
unlikely to build another form which would re-use these elements.