Fix table link :active style in Safari
Note: this is not to do with GOVUK Frontend 3.

Safari is the only browser that doesn't show focus
styles when a link is activated. That means on
other browsers you will see the focus style when
clicking it with a cursor but not on Safari.

Links that appear in our tables (visually) expand
their focus style to fill their table cell. This
is done by styling a :before psuedo-element. We
attached these styles to its :active and :focus
states to mimic most browsers. On Safari, this
produces an issue because these styles were being
applied but the link styles, which show on focus
only, were not. The result for links clicked with
a cursor is a link with :active styling and an
expanded area that looks focused.

This fixes the issue by only applying the expanded
focus styling to the :focus state.