Fix highlight textbox text size on mobile
Our highlight textboxes (textareas, enhanced to
highlight placeholder text) don't scale the size
of their text down for smaller screens. The rest
of the text does (including normal text inputs) so
they look odd.

This could be fixed by adding styles to set just
the size for smaller screens but this would
hard-code the sizes, which are really defined by
design system. This instead uses the govuk-font
helper which will add a few duplicate styles but
will ensure consistency with the rest of our text
in the future.