Remove focus hacks for components
Removes the govuk-frontend/focus folder. We added
its Sass to give our version 2 of GOVUK Frontend
the focus style for version 3.

We're now moving to version 3 so these styles are
no longer needed.

This commit also includes a change to the maps CSS
to remove a variable similarly not needed due to
our moving to version 3.