Update test for preloaded fonts
Before version 3, GOVUK Frontend used a cut of the
New Transport ('nta') font which kept the tabular
numbers version (where each character occupies the
same amount of space, so they line up vertically)
in a separate file. Because the 'light' (normal
weight) and 'bold' variants are in separate files,
this means we needed to load 4 files in total.

Note: the old font also came with .eot and .woff
variants (most users use the .woff2 one), both in
'light' and 'bold', so you
would see 12 files in `app/static`.

Version 3 comes with a new cut of the font, called
'GDS Transport' which has the tabular numbers
version included. We therefore only need to load 2
files, one for 'light', one for 'bold'.

Note: we still serve .woff variants so you there
will now be 4 files in `app/static`.

As reference, you can see all the files for the 2
versions of the font here:

- [GOVUK Frontend v2](https://github.com/alphagov/govuk-frontend/tree/v2.13.0/package/assets/fonts)
- [GOVUK Frontend v3+](https://github.com/alphagov/govuk-frontend/tree/v3.0.0/package/govuk/assets/fonts)