Forward limited headers from template-preview
We have experienced issues locally with template-preview responses to
the admin app because they include a Transfer-Encoding:chunked header
that indicates the response is a streaming response. By including these
in the response from admin to the browser, the browser gets confused,
because it receives a standard non-streamed HTTP response but has a
header telling it to expect chunks. This leads to issues and doesn't
render the template preview correctly in the iframe.

Let's only forward specific headers (eg the Content-Type) and allow
flask on the admin app to set the rest as it sees fit.