Add endpoint to show all letter branding for org
This adds a new endpoint,
`/organisations/<uuid:org_id>/settings/letter-branding`, which will
eventually match the equivalent endpoint for email branding and be used
to show a preview of every brand with the option to make various brands
the default or remove them. For now, it just displays all branding in
the pool and there are no actions possible.