vendor govuk frontend jinja v1.5.1 checkboxes
in https://github.com/alphagov/notifications-admin/pull/3535 we copied
across the source for the govuk frontend njk checkboxes so we could make
some required changes that weren't exposed using the normal macro
parameters.

we'll need to do the same with the new jinja library's version - pull in
the source code, then go through and see which changes we need to make.

This file is pulled from here:

https://github.com/LandRegistry/govuk-frontend-jinja/blob/1.5.1/govuk_frontend_jinja/templates/components/checkboxes/macro.html

Copy across the comment describing changes from the old njk template so
we can start to migrate those changes