Change colour used to replace grey-3
Replaces the colour chosed originally (#e0e1e2)
with one that's a little warmer (#e2e1e0), to
better match the footer. This was suggested in a
comment by @quis on the associated pull request:

https://github.com/alphagov/notifications-admin/pull/4405#issuecomment-1300121106

We took that colour and managed to generate it
from the footer colour by using the `govuk-shade`
function.