Use form legends as page title
When the form legend is the same as the page title it’s not necessary to
have both on the page. The proper way to do this is through a
`param_extensions` argument.

This also fixes the spacing to make these legends-as-page-headings have
the same distance from the back link as other `<h1>`s