remove formGroup=False from govuk checkbox overrides
In https://github.com/alphagov/notifications-admin/pull/3654 we added an
override to remove the govuk-form-group class if we passed formGroup =
false in to the njk macro. We did this so we could remove a bottom
margin. However, adding a separate override class is consistent with how
we've modified margins elsewhere, and is built in to the original govuk
frontend macro.

By simplifying the njk macro we'll make it easier to compare the jinja
and njk macros as we migrate.