copy overrides from old njk checkbox to new jinja macro
item.classes - moved across without any modifications

item.hint.classes - introduced in govuk frontend 3.5.0, so manual
changes to the template are required any more.

asList - The `item.conditional.html` section is the same content
regardless of params.asList, but is just a child of a different element
depending on whether you're in a list or in a div, so we've moved that
definition up to the top to avoid duplication.

item.children - moved across without any modifications

item.before and item.after - moved across without any modifications.
these weren't mentioned in the original comment at the top of the file
but aren't present in the original source njk.