Fix message when service has no letter branding
This was broken by 245f1b9524aba8aeff540f056d0cbd44aaf43dfe
because the default value for letter branding name was changed from
`'no branding'` to `None`.

This commit fixes the bug by checking for a `None` value in the template
and makes sure the message has test coverage.