In template folder js find button based on value, not type
Ffrom govuk frontend v3.1.0 onwards, buttons have no type unless
overriden (previously they had type=submit). If a button is inside a
<form> element, then the type is implicitly submit anyway[1].

We can no longer use `[type=submit]` as a selector to find the button
from within the complex template folder js. Since we've got the value
selector in the addCancelButton it's cleaner and tidier to use that
anyway.

[^1]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#attr-type