Add better example text
Lorem ipsum text was fine when it was just our team looking at this
page.

Now that we show this page to users who are previewing their own
branding, we should make it clearer that the content is an example.

Only the first 3 lines of the content are visible in the `iframe`, but I
figure it’s still useful to have a page which quickly previews some of
the formatting too.