Increase height of placeholder redaction mask
The GDS Transport font, which we moved to
recently, sits slightly lower on the line which
seems to have resulted in the text we display for
redacted placeholders poking out from below the
bottom of the mask we cover it with.

This increases the height of the mask so it covers
its text again.

Note: We also replace the text of any placeholders
with 'hidden' so this issue will not have
disclosed any information set as redacted.