move service_navigation from {% include %} to {% import %}
service_navigation.html contained both macros and some html. When
calling `{% include service_navigation.html %}`, we would execute the
module-level template context of that file, with the outer file's
context (The context contains things like current_service).

This leads to surprising things. When you try and import one of the
macros from that file, it would execute that module template code.
However, the jinja `{% from x import y %}` command _doesn't include
context_ (unless you specifically say `{% from x import y with context
%}`. Rather than pass that context in just so that some code that we
don't even render doesnt have issues,

This isn't an issue currently but if we move from ChainedUndefined to
Undefined, it'll complain that `current_service` isn't set as the
context is empty in that nested template render call.

To solve this, move everything in the file into a macro. That way, when
we import, nothing gets executed. This also requires we pass in the
variables, as mentioned in the jinja context docs.

https://flask.palletsprojects.com/en/2.2.x/templating/#standard-context

Read "The Jinja Context Behavior" for a little more context, heh.